Added FPS controller for testing. Made UI panels draggable.
Moved the main camera into a Subject object, and added a FPS controller to the Subject so that I can move the subject camera without having a VR setup while testing.

Added a script to allow UI elements in the Experimenter UI to be dragged when you click on the panel title.