Adjustments to proxemics to address new VR system
Each skin now has a AgentSkin component.
Each skin now has a gaze target empty parented to its head bone, and accessible through the Agent skin.
AgentSkinController is now AgentController, and provides access to the current gaze target.
Moved reticle out to top level of hierarchy so that it will still draw to the display.
Tested having a VR reticle, but this presents issues with 2 eye focus.
Updated draggable panel to work with old and new input systems.

Added a small UI to the VR view to monitor the gaze and distance data.
Currently gaze is drawing rays in the scene view to show what vectors are.

The gaze reading probably is too sensitive and needs to be given a wider reading when close.

The agent head is aiming down at the aim constraint for some reason. Not sure why.
Also looks like aiming object has to be under the rig, which is not great.