Added camera and FPS view controller for the experimenter.
Added a new camera to provide the desktop view as opposed to the headset view.
Created an ExperimenterViewController with an attached FPS system that can be turned on and off. New camera is its child.
Added a new Input System control to take in mouse input for the FPS

Added a Conditional Hide Property and Panel for hiding elements in the inspector.