Documentation Update
Made a pass through all the source code files to update the comments and documentation. No real code changes other than simple cleanup.

Also noticed that moving back and forth from Windows to Mac is causing git to see a lot of file changes when nothing changed. I think the issue is windows not understanding the unix file permissions. Only the CS files should have updated with this commit. Will need to look into the git settings to fix this.