Added Agent Model
Created AgentModel class to model the arrangement of the agent model relationships in memory. This is now populated with all the parts when the model is loaded.

AgentSkin now updates arm IK positions to keep them from moving when turned on and off.
AgentSkin now listens for keyboard input to turn on and off various IKs

InputControls and InputManager - C# events / listener pattern
- centralized keyboard input to a single input manager.
- maps for each set of controls defined in InputControls.
- maps can be switched on and off from the manager. This will also trigger an event others can listen for.
- manager is accessible from the AFManager singleton.
>> updated ExperimenterViewController to use
>> updated FPSController to use
>> updated AgentSkin to use

Coroutine Animation
- Learned how to do lerp animation and run it from a coroutine.
- Created a class of easing functions to use with Lerps

Replaced the Woman avatar prefab to fix some of the animation stuff coming in.

Added AppUtilities class to hold some useful functions for a depth first traversal of GameObject tree.

TODO: