Fixed Desktop/VR Input Issue
We have been pushing VR input to display 1 and desktop to display 2, but this seems to be an issue. Setting all to display 1 and just using the camera depth to put the desktop UI at top seems to work.
For UI elements:
Make sure Canvas objects are set to send canvas to display 1. These were on 2. So far I have only updated the experimenter UI.