Subject POV for Experimenter View
Added a small POV view to the experimenter view showing what the subject is seeing. Buttons are included to toggle the view open and closed.

Added a secondary camera to the subject to capture the POV. This feeds to a render texture thats displayed by the panel in the UI.