Fixed overlapping triggers for expressions and head tilts
If the head or face were already in a pose or expression, hitting that trigger again would leave a trigger set but not cleared, and cause a glitch on the next animation triggered.
AgentSkin now keeps track of the current head tilt and the current expression. If a call comes in to execute from the current state, then it's ignored, avoiding this issue.
The AnimationControllers for both models now also transition between tilted one side to the other side correctly.