Hand constraint pausing on sit shift
If the hand constraints are on, then they will pause when the sit shift animation is activated, then resume.

Added pasue constraint behavior to the animation controllers for sit shift animations.
Set AgentSkin to record and resume constraints when fired.
Set hand constraints to move to current position when activated.