Added correct global average logging, logging cleanup
A lot of logging was not being done in the results file, because I took several of the headers out by mistake. So these are now added back in.

The results file was logging averages for data, but it was just the last interval average. Added code and algo to calculate a global average and log this data into the results file at the end of trial.

Adjusted StartUI to properly collect and validate the data so it ends up in the results file correctly.

Changed PPID to hold ID_NAME.
need to get the results folder to be the whole PPID and not just the number part.