Fixed File Save Bug
Turns out that the file save location was not being passed to the file saver.
Now when the begin button is pressed, the system reads through all local file savers and sets the correct file path.
Also removed the confusing filepath var in SettingUIController. We will just use the text value of the UI field.