Bug fixes in prompter panel
Input controls now properly only come on when the adjustment panel is displayed, not every time the prompter is displayed. This was conflicting with typing in the chat box.
Prompter now waits until it has finished fading out to clear the text.

Can not turn input controls on and off independent of visibility.
Can now pass a callback to the fade function of panel
Removed unused toggle functions.